14. Przeorganizowanie projektu i usuniecie pliku Node
