+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -1,6 +1,37 @@
 package edu.kis.vh.nursery.list;
 
 public class IntLinkedList {
+    private static class Node {
+
+        private final int VALUE;
+        private Node prev;
+        private Node next;
+
+        public Node(int i) {
+            VALUE = i;
+        }
+
+        public int getVALUE() {
+            return VALUE;
+        }
+
+        public Node getPrev() {
+            return prev;
+        }
+
+        public void setPrev(Node prev) {
+            this.prev = prev;
+        }
+
+        public Node getNext() {
+            return next;
+        }
+
+        public void setNext(Node next) {
+            this.next = next;
+        }
+    }
+
 
     private static final int IF_EMPTY = -1;
     private Node last;
